(feat) Add api testing routing
Also refactored api handlers to folder api_handlers.
