--- mergeTest.txt
@@ -1,2 +1,0 @@
-"Screw your merge"
-SUP BACKSTR33T BOYZ
